Removed unnecessary files and prepered build_group_20 and group_20
