--- .github/workflows/parse_submissions.yaml
@@ -1,67 +1,0 @@
-name: Deploy operating systems
-
-
-on:
-  # Runs on pushes targeting the default branch
-  workflow_run:
-    workflows: [ "Operating Systems Pipeline" ]
-    branches: [ "master" ]
-    types:
-      - completed
-
-  # Allows you to run this workflow manually from the Actions tab
-  workflow_dispatch:
-
-# Sets permissions of the GITHUB_TOKEN to allow deployment to GitHub Pages
-permissions:
-  contents: read
-  pages: write
-  id-token: write
-
-# Allow one concurrent deployment
-concurrency:
-  group: "pages"
-  cancel-in-progress: true
-
-jobs:
-  # Build job
-  build:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v3
-
-      - name: Install python
-        uses: actions/setup-python@v4
-        with:
-          python-version: '3.10'
-
-
-      - name: Install dependencies
-        run: python3 -m pip install ./web
-
-
-      - name: Build with Python and Jinja2
-        run: mkdir -p ./_site && python3 ./web/web/compile.py > ./_site/index.html
-
-      - name: Copy static files to _site
-        run: cp -r ./web/web/static/* ./_site
-
-      - name: Setup Pages
-        id: pages
-        uses: actions/configure-pages@v2
-
-      - name: Upload artifact
-        uses: actions/upload-pages-artifact@v1
-
-  # Deployment job
-  deploy:
-    environment:
-      name: github-pages
-      url: ${{ steps.deployment.outputs.page_url }}
-    runs-on: ubuntu-latest
-    needs: build
-    steps:
-      - name: Deploy to GitHub Pages
-        id: deployment
-        uses: actions/deploy-pages@v1
