+++ .vscode/settings.json
@@ -1,6 +1,6 @@
 {
-    "cmake.buildDirectory": "${workspaceFolder}/build_group_name",
-    "cmake.sourceDirectory": "${workspaceFolder}/group_name",
+    "cmake.buildDirectory": "${workspaceFolder}/build_group_20",
+    "cmake.sourceDirectory": "/home/amrapstad/ikt218-osdev/group_20",
     "C_Cpp.errorSquiggles": "disabled"
 
 }
