+++ group_20/grub/Dockerfile
@@ -1,0 +1,4 @@
+FROM ubuntu:22.04
+RUN apt-get update && apt-get install grub-pc-bin grub-common parted xorriso udev -y
+WORKDIR /
+CMD ["/bin/bash"]
