+++ group_20/grub/build_docker.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+docker buildx build . --platform=linux/amd64 -t perara/grub
