+++ group_20/grub/grub.cfg.in
@@ -1,0 +1,7 @@
+set timeout=0
+set default=0
+
+menuentry "${OS_NAME}" {
+	${GRUB_MULTIBOOT_VERSION} /boot/kernel.bin
+    boot
+}
