+++ group_20/src/stdlib/src/system.c
@@ -1,0 +1,1 @@
+#include "system.h"
