--- group_name/grub/Dockerfile
@@ -1,4 +1,0 @@
-FROM ubuntu:22.04
-RUN apt-get update && apt-get install grub-pc-bin grub-common parted xorriso udev -y
-WORKDIR /
-CMD ["/bin/bash"]
