--- group_name/grub/build_docker.sh
@@ -1,2 +1,0 @@
-#!/bin/bash
-docker buildx build . --platform=linux/amd64 -t perara/grub
