--- group_name/src/kernel/src/main.cpp
@@ -1,12 +1,0 @@
-
-#include "system.h"
-
-// Define entry point in asm to prevent C++ mangling
-extern "C"{
-    void kernel_main();
-}
-
-void kernel_main()
-{
-
-}
