--- group_name/src/stdlib/src/system.c
@@ -1,1 +1,0 @@
-#include "system.h"
